Corección en Tipos de Datos y Conversiones (Consumo desde EF CORE TOOLS)

+++ Entidades/Cuatrimestre.cs
@@ -13,7 +13,7 @@ namespace Entidades
             GrupoCuatrimestre = new HashSet<GrupoCuatrimestre>();
         }
 
-        public int IdCuatrimestre { get; set; }
+        public short IdCuatrimestre { get; set; }
         public string Periodo { get; set; }
         public int Anio { get; set; }
         public DateTime Inicio { get; set; }
